Set graph as inactive before deletion #3574
To give time to allow all resources to be removed, set the graph as
inactive before starting.

+++ arches/app/views/graph.py
@@ -247,7 +247,10 @@ class GraphManagerView(GraphBaseView):
     def delete(self, request, graphid):
         try:
             graph = Graph.objects.get(graphid=graphid)
-            graph.delete_instances()
+            if graph.isresource:
+                graph.isactive = False
+                graph.save()
+                graph.delete_instances()
             graph.delete()
             return JSONResponse({'success':True})
         except GraphValidationError as e:
